refactor: Extract users api into rest-api-client package
