+++ packages/frontend/editor-ui/src/__tests__/server/factories/user.ts
@@ -1,7 +1,7 @@
 import { Factory } from 'miragejs';
 import { faker } from '@faker-js/faker';
 import { SignInType } from '@/constants';
-import type { IUser } from '@/Interface';
+import type { IUser } from '@n8n/rest-api-client/api/users';
 
 export const userFactory = Factory.extend<IUser>({
 	id(i: number) {
