+++ packages/frontend/editor-ui/src/__tests__/server/models/user.ts
@@ -1,4 +1,4 @@
-import type { IUser } from '@/Interface';
+import type { IUser } from '@n8n/rest-api-client/api/users';
 import { Model } from 'miragejs';
 import type { ModelDefinition } from 'miragejs/-types';
