+++ packages/frontend/editor-ui/src/api/invitation.ts
@@ -1,4 +1,5 @@
-import type { CurrentUserResponse, IInviteResponse, InvitableRoleName } from '@/Interface';
+import type { IInviteResponse, InvitableRoleName } from '@/Interface';
+import type { CurrentUserResponse } from '@n8n/rest-api-client/api/users';
 import type { IRestApiContext } from '@n8n/rest-api-client';
 import type { IDataObject } from 'n8n-workflow';
 import { makeRestApiRequest } from '@n8n/rest-api-client';
