+++ packages/frontend/editor-ui/src/api/workflow-webhooks.ts
@@ -1,4 +1,4 @@
-import type { IUser } from '@/Interface';
+import type { IUser } from '@n8n/rest-api-client/api/users';
 import { post } from '@n8n/rest-api-client';
 
 const N8N_API_BASE_URL = 'https://api.n8n.io/api';
