+++ packages/frontend/editor-ui/src/components/InviteUsersModal.vue
@@ -6,9 +6,9 @@ import type {
 	FormFieldValueUpdate,
 	IFormInputs,
 	IInviteResponse,
-	IUser,
 	InvitableRoleName,
 } from '@/Interface';
+import type { IUser } from '@n8n/rest-api-client/api/users';
 import { EnterpriseEditionFeature, VALID_EMAIL_REGEX, INVITE_USER_MODAL_KEY } from '@/constants';
 import { ROLE } from '@n8n/api-types';
 import { useUsersStore } from '@/stores/users.store';
