+++ packages/frontend/editor-ui/src/components/SettingsUsers/SettingsUsersActionsCell.test.ts
@@ -6,7 +6,7 @@ import { ROLE, type UsersList } from '@n8n/api-types';
 import { type UserAction } from '@n8n/design-system';
 import SettingsUsersActionsCell from '@/components/SettingsUsers/SettingsUsersActionsCell.vue';
 import { createComponentRenderer } from '@/__tests__/render';
-import type { IUser } from '@/Interface';
+import type { IUser } from '@n8n/rest-api-client/api/users';
 
 const baseUser: UsersList['items'][number] = {
 	id: '1',
