+++ packages/frontend/editor-ui/src/components/SettingsUsers/SettingsUsersTable.test.ts
@@ -7,7 +7,7 @@ import { type UserAction } from '@n8n/design-system';
 import SettingsUsersTable from '@/components/SettingsUsers/SettingsUsersTable.vue';
 import { createComponentRenderer } from '@/__tests__/render';
 import { useEmitters } from '@/__tests__/utils';
-import type { IUser } from '@/Interface';
+import type { IUser } from '@n8n/rest-api-client/api/users';
 
 const { emitters, addEmitter } = useEmitters<
 	'settingsUsersRoleCell' | 'settingsUsersActionsCell' | 'n8nDataTableServer'
