+++ packages/frontend/editor-ui/src/components/banners/V1Banner.test.ts
@@ -3,7 +3,7 @@ import V1Banner from './V1Banner.vue';
 import { createPinia, setActivePinia } from 'pinia';
 import { useUsersStore } from '@/stores/users.store';
 import { ROLE } from '@n8n/api-types';
-import type { IUser } from '@/Interface';
+import type { IUser } from '@n8n/rest-api-client/api/users';
 
 const renderComponent = createComponentRenderer(V1Banner, {
 	global: {
