+++ packages/frontend/editor-ui/src/composables/useCalloutHelpers.ts
@@ -6,7 +6,7 @@ import { usePostHog } from '@/stores/posthog.store';
 import { useUsersStore } from '@/stores/users.store';
 import { RAG_STARTER_WORKFLOW_EXPERIMENT, VIEWS } from '@/constants';
 import { getRagStarterWorkflowJson } from '@/utils/easyAiWorkflowUtils';
-import { updateCurrentUserSettings } from '@/api/users';
+import { updateCurrentUserSettings } from '@n8n/rest-api-client/api/users';
 import { useWorkflowsStore } from '@/stores/workflows.store';
 
 export function useCalloutHelpers() {
