+++ packages/frontend/editor-ui/src/features/dataStore/DataStoreView.vue
@@ -10,7 +10,8 @@ import { computed, onMounted, ref } from 'vue';
 import { useRoute } from 'vue-router';
 import { ProjectTypes } from '@/types/projects.types';
 import { useProjectsStore } from '@/stores/projects.store';
-import type { IUser, SortingAndPaginationUpdates, UserAction } from '@/Interface';
+import type { SortingAndPaginationUpdates, UserAction } from '@/Interface';
+import type { IUser } from '@n8n/rest-api-client/api/users';
 import type { DataStoreResource } from '@/features/dataStore/types';
 import DataStoreCard from '@/features/dataStore/components/DataStoreCard.vue';
 import { useSourceControlStore } from '@/stores/sourceControl.store';
