+++ packages/frontend/editor-ui/src/features/dataStore/components/DataStoreCard.test.ts
@@ -2,7 +2,8 @@ import { createComponentRenderer } from '@/__tests__/render';
 import DataStoreCard from './DataStoreCard.vue';
 import { createPinia, setActivePinia } from 'pinia';
 import type { DataStoreResource } from '@/features/dataStore/types';
-import type { UserAction, IUser } from '@/Interface';
+import type { UserAction } from '@/Interface';
+import type { IUser } from '@n8n/rest-api-client/api/users';
 
 vi.mock('vue-router', () => {
 	const push = vi.fn();
