+++ packages/frontend/editor-ui/src/features/insights/insights.store.test.ts
@@ -6,7 +6,7 @@ import { useRootStore } from '@n8n/stores/useRootStore';
 import { useUsersStore } from '@/stores/users.store';
 import { useSettingsStore } from '@/stores/settings.store';
 import { mockedStore, type MockedStore } from '@/__tests__/utils';
-import type { IUser } from '@/Interface';
+import type { IUser } from '@n8n/rest-api-client/api/users';
 import { reactive } from 'vue';
 import type { FrontendModuleSettings } from '@n8n/api-types';
