+++ packages/frontend/editor-ui/src/stores/users.store.ts
@@ -8,18 +8,18 @@ import {
 	ROLE,
 	type UsersListFilterDto,
 } from '@n8n/api-types';
-import type { UpdateGlobalRolePayload } from '@/api/users';
-import * as usersApi from '@/api/users';
+import type { UpdateGlobalRolePayload } from '@n8n/rest-api-client/api/users';
+import * as usersApi from '@n8n/rest-api-client/api/users';
 import { BROWSER_ID_STORAGE_KEY } from '@n8n/constants';
 import { PERSONALIZATION_MODAL_KEY } from '@/constants';
 import { STORES } from '@n8n/stores';
+import type { InvitableRoleName } from '@/Interface';
+import type { IUserResponse } from '@n8n/rest-api-client/api/users';
 import type {
-	IPersonalizationLatestVersion,
 	IUser,
-	IUserResponse,
 	CurrentUserResponse,
-	InvitableRoleName,
-} from '@/Interface';
+	IPersonalizationLatestVersion,
+} from '@n8n/rest-api-client/api/users';
 import { getPersonalizedNodeTypes } from '@/utils/userUtils';
 import { defineStore } from 'pinia';
 import { useRootStore } from '@n8n/stores/useRootStore';
