+++ packages/frontend/editor-ui/src/types/externalHooks.ts
@@ -15,10 +15,10 @@ import type {
 	INodeCreateElement,
 	INodeUi,
 	INodeUpdatePropertiesInformation,
-	IPersonalizationLatestVersion,
 	IWorkflowDb,
 	NodeFilterType,
 } from '@/Interface';
+import type { IPersonalizationLatestVersion } from '@n8n/rest-api-client/api/users';
 import type { IWorkflowTemplateNode } from '@n8n/rest-api-client/api/templates';
 import type { ComponentPublicInstance } from 'vue';
 import type { useWebhooksStore } from '@/stores/webhooks.store';
