+++ packages/frontend/editor-ui/src/utils/rbac/middleware/role.test.ts
@@ -1,6 +1,6 @@
 import { roleMiddleware } from '@/utils/rbac/middleware/role';
 import { useUsersStore } from '@/stores/users.store';
-import type { IUser } from '@/Interface';
+import type { IUser } from '@n8n/rest-api-client/api/users';
 import type { RouteLocationNormalized } from 'vue-router';
 import { ROLE } from '@n8n/api-types';
 import { VIEWS } from '@/constants';
