+++ packages/frontend/editor-ui/src/utils/userUtils.ts
@@ -67,8 +67,8 @@ import type {
 	IPersonalizationSurveyAnswersV3,
 	IPersonalizationSurveyVersions,
 	IUser,
-	ILogInStatus,
-} from '@/Interface';
+} from '@n8n/rest-api-client/api/users';
+import type { ILogInStatus } from '@/Interface';
 import type { IPersonalizationSurveyAnswersV4 } from 'n8n-workflow';
 
 /*
